requirements.txt
# Core Dependencies
kafka-python==2.0.2
pyspark==3.3.0
psycopg2-binary==2.9.3
pandas==1.5.0
numpy==1.26.4
python-dotenv==0.21.0
Faker==19.3.0
kaggle==1.5.13

# Testing Dependencies
pytest==7.1.3
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-env==0.8.1
coverage==7.2.3

# Development Dependencies
black==22.3.0
flake8==6.0.0
mypy==1.3.0
isort==5.12.0
